Implement /delivery/register_consumer API method
Summary:
Declare a queue, bind the queue to an exchange.

A first approach was to connect to the broker at server start,
a channel, then to interact with the channel when requests come.

This approach isn't optimal as:
  - channel can be closed by errors like a non existing exchange
  - we want to recover if the broker restarts
  - the pika loop and the Werkzeug loop could behave differently

So each request will open/close a connection to the broker.

Ref T1210

Test Plan: Fire an API request, get a key, check a queue exists matching this key.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1210

Differential Revision: https://devcentral.nasqueron.org/D991